neural_network/chatbot/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.9
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
eval_type_backport==0.2.0
filelock==3.16.1
frozenlist==1.4.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nodeenv==1.9.1
numpy==2.1.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pre_commit==4.0.1
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
ruff==0.7.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.1
tabulate==0.9.0
together==1.3.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
yarl==1.13.1
